Safari-609.1.11.1/LayoutTests/http/tests/loading/slow-parsing-subframe.html

<html>
<script>
    if (window.testRunner) {
      testRunner.dumpAsText();
      testRunner.waitUntilDone();
    }
</script>
<body>
This tests that we get frame load callbacks in the proper order when a subframe
is parsed and created from within a timer callback.
<script>
  // Force a sleep for 1 second, to make sure the rest of the parsing is done
  // via a timer.
  var start = new Date()
  var startMS = start.getTime()
  while (true) {
    var now = new Date()
    if (now.getTime() - startMS > 1000) {
      break;
    }
  }
</script>
<iframe name="f1" src="http://127.0.0.1:8000/misc/resources/200.html"></iframe><br>
</body>
</html>
